requirements/common.txt
# Core Stuff
# -------------------------------------
Django==2.0.1

# WSGI HTTP Server
# -------------------------------------
uWSGI==2.0.15

# Configuration
# -------------------------------------
argon2-cffi==18.1.0
django-environ==0.4.4
django-sites==0.10
python-dotenv==0.7.1
django-cors-headers==2.1.0

# Staticfiles
# -------------------------------------
whitenoise==3.3.1

# Extensions
# -------------------------------------
pytz==2017.3

# Models
# -------------------------------------
psycopg2==2.7.3.2

Pillow==5.0.0
django-extensions==1.9.8
django-uuid-upload-path==1.0.0
django-versatileimagefield==1.8.1  # versatile image field degrade Pillow version to 4.0.0

# REST APIs
# -------------------------------------
djangorestframework==3.7.7
django-rest-swagger==2.1.2

# Static files
# -------------------------------------
django-compressor==2.2
django-libsass==0.7
django-compressor-autoprefixer==0.1.0

# LOGGING
# -------------------------------------
django-log-request-id==1.3.2

# Documentation
# -------------------------------------
mkdocs==0.17.2



# Auth Stuff
# -------------------------------------
PyJWT==1.5.3
